backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
httpx>=0.25.0
python-jose>=3.3.0
python-multipart>=0.0.6
supabase>=1.0.3
prisma>=0.10.0
celery>=5.3.4
redis>=5.0.1
PyJWT>=2.8.0
pytest>=7.4.3
httpx>=0.25.0
python-dotenv>=1.0.0
